src/requirements.txt
Pygame.pip
Image download of Scene.png for the image in video. Can use any image you desire by replacing "scene.png" with the new image file in line 10
